refactor: adjust file structure for jest/ts
