+++ tsconfig.json
@@ -17,10 +17,10 @@
     "pretty": true,
     "removeComments": true,
     "resolveJsonModule": true,
-    "rootDir": "psi_cardinality/javascript",
+    "rootDir": "psi_cardinality/javascript/src",
     "sourceMap": true,
     "strict": true,
     "target": "es2020"
   },
-  "include": ["psi_cardinality/javascript/**/*"]
+  "include": ["psi_cardinality/javascript/src/**/*"]
 }
